+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Added the alarm label to the upcoming alarm notification ([#183])
 
 ## [1.6.0] - 2026-01-30
 ### Added
@@ -106,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#107]: https://github.com/FossifyOrg/Clock/issues/107
 [#144]: https://github.com/FossifyOrg/Clock/issues/144
 [#158]: https://github.com/FossifyOrg/Clock/issues/158
+[#183]: https://github.com/FossifyOrg/Clock/issues/183
 [#206]: https://github.com/FossifyOrg/Clock/issues/206
 [#207]: https://github.com/FossifyOrg/Clock/issues/207
 [#247]: https://github.com/FossifyOrg/Clock/issues/247